+++ packages/article/article-meta.proptypes.js
@@ -1,0 +1,24 @@
+import PropTypes from "prop-types";
+
+const TextNode = PropTypes.shape({ text: PropTypes.string });
+
+const nodeShape = {
+  name: PropTypes.string.isRequired,
+  attributes: PropTypes.object.isRequired
+};
+
+nodeShape.children = PropTypes.arrayOf(
+  PropTypes.oneOfType([PropTypes.shape(nodeShape), TextNode])
+).isRequired;
+
+export const articleMetaPropTypes = {
+  byline: PropTypes.arrayOf(PropTypes.shape(nodeShape)),
+  publishedTime: PropTypes.string,
+  publicationName: PropTypes.string
+};
+
+export const articleMetaDefaultPropTypes = {
+  byline: [],
+  publishedTime: null,
+  publicationName: null
+};