+++ packages/link/link.web.js
@@ -2,7 +2,7 @@ import React from "react";
 import propTypes from "./link.proptypes";
 
 const Link = ({ url, onPress, children }) => (
-  <a href={url} onClick={onPress}>
+  <a href={url} onClick={onPress} style={{ textDecoration: "none" }}>
     {children}
   </a>
 );