+++ packages/markup/tree-proptype.js
@@ -1,15 +1,11 @@
 import PropTypes from "prop-types";
 
-const TextNode = PropTypes.shape({ text: PropTypes.string });
-
 const nodeShape = {
   name: PropTypes.string.isRequired,
   attributes: PropTypes.object.isRequired
 };
 
-nodeShape.children = PropTypes.arrayOf(
-  PropTypes.oneOfType([PropTypes.shape(nodeShape), TextNode])
-).isRequired;
+nodeShape.children = PropTypes.arrayOf(PropTypes.shape(nodeShape)).isRequired;
 
 const Node = PropTypes.shape(nodeShape);