+++ packages/article-byline/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.5.7"></a>
+## [0.5.7](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.5.6...@times-components/article-byline@0.5.7) (2017-09-28)
+
+
+
+
+<a name="0.5.6"></a>
+## [0.5.6](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.5.5...@times-components/article-byline@0.5.6) (2017-09-26)
+
+
+
+
+<a name="0.5.5"></a>
+## [0.5.5](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.5.4...@times-components/article-byline@0.5.5) (2017-09-25)
+
+
+### Bug Fixes
+
+* package.json main ([ed150b8](https://github.com/newsuk/times-components/commit/ed150b8))
+
+
+
+
+<a name="0.5.4"></a>
+## [0.5.4](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.5.2...@times-components/article-byline@0.5.4) (2017-09-25)
+
+
+
+
+<a name="0.5.3"></a>
+## [0.5.3](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.5.2...@times-components/article-byline@0.5.3) (2017-09-25)
+
+
+
+
 <a name="0.5.2"></a>
 ## [0.5.2](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.5.1...@times-components/article-byline@0.5.2) (2017-09-25)