+++ packages/card/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.6.5"></a>
+## [0.6.5](https://github.com/newsuk/times-components/compare/@times-components/card@0.6.4...@times-components/card@0.6.5) (2017-09-28)
+
+
+
+
+<a name="0.6.4"></a>
+## [0.6.4](https://github.com/newsuk/times-components/compare/@times-components/card@0.6.3...@times-components/card@0.6.4) (2017-09-26)
+
+
+
+
+<a name="0.6.3"></a>
+## [0.6.3](https://github.com/newsuk/times-components/compare/@times-components/card@0.6.2...@times-components/card@0.6.3) (2017-09-25)
+
+
+### Bug Fixes
+
+* package.json main ([ed150b8](https://github.com/newsuk/times-components/commit/ed150b8))
+
+
+
+
 <a name="0.6.2"></a>
 ## [0.6.2](https://github.com/newsuk/times-components/compare/@times-components/card@0.6.1...@times-components/card@0.6.2) (2017-09-25)