+++ packages/image/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="1.4.1"></a>
+## [1.4.1](https://github.com/newsuk/times-components/compare/@times-components/image@1.4.0...@times-components/image@1.4.1) (2017-09-25)
+
+
+### Bug Fixes
+
+* package.json main ([ed150b8](https://github.com/newsuk/times-components/commit/ed150b8))
+
+
+
+
 <a name="1.4.0"></a>
 # [1.4.0](https://github.com/newsuk/times-components/compare/@times-components/image@1.2.1...@times-components/image@1.4.0) (2017-09-22)